Updated android branch support
