+++ packages/config/src/android/Branch.ts
@@ -1,5 +1,10 @@
 import { ExpoConfig } from '../Config.types';
-import { Document, getMainApplication } from './Manifest';
+import {
+  addMetaDataItemToMainApplication,
+  Document,
+  getMainApplication,
+  removeMetaDataItemFromMainApplication,
+} from './Manifest';
 
 export function getBranchApiKey(config: ExpoConfig) {
   return config.android?.config?.branch?.apiKey ?? null;
@@ -8,34 +13,14 @@ export function getBranchApiKey(config: ExpoConfig) {
 export async function setBranchApiKey(config: ExpoConfig, manifestDocument: Document) {
   const apiKey = getBranchApiKey(config);
 
-  if (!apiKey) {
-    return manifestDocument;
-  }
+  const mainApplication = getMainApplication(manifestDocument);
 
-  let mainApplication = getMainApplication(manifestDocument);
-
-  let existingBranchApiKeyItem;
-  const newBranchApiKeyItem = {
-    $: {
-      'android:name': 'io.branch.sdk.BranchKey',
-      'android:value': apiKey,
-    },
-  };
-  if (mainApplication?.['meta-data']) {
-    existingBranchApiKeyItem = mainApplication['meta-data'].filter(
-      (e: any) => e['$']['android:name'] === 'io.branch.sdk.BranchKey'
-    );
-    if (existingBranchApiKeyItem.length) {
-      existingBranchApiKeyItem[0]['$']['android:value'] = apiKey;
-    } else {
-      mainApplication['meta-data'].push(newBranchApiKeyItem);
-    }
+  if (apiKey) {
+    // If the item exists, add it back
+    addMetaDataItemToMainApplication(mainApplication, 'io.branch.sdk.BranchKey', apiKey);
   } else {
-    if (!mainApplication) {
-      mainApplication = { $: { 'android:name': '.MainApplication' } };
-    }
-    mainApplication['meta-data'] = [newBranchApiKeyItem];
+    // Remove any existing item
+    removeMetaDataItemFromMainApplication(mainApplication, 'io.branch.sdk.BranchKey');
   }
-
   return manifestDocument;
 }
